[AN-87] 받은 카드를 저장할 카테고리를 선택한다.
Former-commit-id: 86fb287093f4e45804c04280f0b055b6d6745875

+++ app/src/main/java/act/sds/samsung/angelman/presentation/activity/ShareCardActivity.java
@@ -167,8 +167,11 @@ public class ShareCardActivity extends AppCompatActivity {
             public void onClick(View v) {
                 if (shareCardModel != null) {
                     try {
-                        FileUtil.unzip(shareFilePath, ContentsUtil.getTempFolder());
                         CategoryModel selectItem = categorySelectDialog.getSelectItem();
+                        if(selectItem == null) {
+                            return;
+                        }
+                        FileUtil.unzip(shareFilePath, ContentsUtil.getTempFolder());
                         CardModel cardModel = saveNewSharedCard(shareCardModel, selectItem.index);
                         ContentsUtil.copySharedFiles(cardModel);
                         FileUtil.removeFilesIn(ContentsUtil.getTempFolder());
